requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
flask_socketio==5.3.6
psycopg2-binary==2.9.9
eventlet==0.35.2
numpy